+++ pkg/NEWS.md
@@ -1,5 +1,8 @@
 # ctv 0.9-3
 
+* Enhance handling of archived packages on CRAN when `read.ctv(..., cran = TRUE)`
+  and `ctv2html(..., cran = TRUE)`.
+
 * Further improvements in `ctv2html()`: Generate HTML5 with `pandoc` now, improve
   preservation of link breaks (`--wrap=preserve`).